+++ magnific_ui/lib/widget/single_async_scope.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 
-typedef SingleAsyncScopeCallback = Future<void> Function(
-  Future<void> Function() onPressed,
-);
-
 /// Builder that doesn't provide a action callback if there's already one being used.
 typedef SingleActionProcessWidgetBuilder = Widget Function(
-  SingleAsyncScopeCallback? singleAsyncScopeCaller,
+  AsyncScopeReference reference,
   Widget? child,
 );
 
-/// A widget builder that provides a [SingleAsyncScopeCallback] to its child
-/// when no other asynchronous operation from a previously invoked [SingleAsyncScopeCallback] is running.
+/// A widget builder that provides a [AsyncScopeReference] to its child.
+/// [AsyncScopeReference] only allows call when no other asynchronous operation from a previously invoked [AsyncScopeReference.call] is running.
 ///
 /// Useful for ignoring tap gestures and show loading indicator
 /// when a button is pressed while processing a future.
@@ -34,32 +30,53 @@ class SingleAsyncScope extends StatefulWidget {
   _SingleAsyncScopeState createState() => _SingleAsyncScopeState();
 }
 
-class _SingleAsyncScopeState extends State<SingleAsyncScope> {
+mixin AsyncScopeReference {
+  bool get isProcessing;
+  bool get isNotProcessing;
+
+  Future<void> call(Future<void> Function() onPressed);
+}
+
+class _SingleAsyncScopeState extends State<SingleAsyncScope>
+    with AsyncScopeReference {
   bool _isProcessing = false;
 
-  Future<void> _onPressed(
+  @override
+  bool get isNotProcessing => !_isProcessing;
+
+  @override
+  bool get isProcessing => _isProcessing;
+
+  @override
+  Future<void> call(
     Future<void> Function() onPressed,
   ) async {
     if (_isProcessing) {
       return;
     }
-    setState(() {
-      _isProcessing = true;
-    });
+    _isProcessing = true;
+    if (mounted) {
+      setState(() {
+        /** update state */
+      });
+    }
     try {
       await onPressed();
     } catch (e, s) {
       widget.onError?.call(e, s);
     }
-    setState(() {
-      _isProcessing = false;
-    });
+    _isProcessing = false;
+    if (mounted) {
+      setState(() {
+        /** update state */
+      });
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     return widget.builder(
-      _isProcessing ? null : _onPressed,
+      this,
       widget.child,
     );
   }