+++ magnific_ui/lib/widget/widget.dart
@@ -4,3 +4,4 @@ export 'image_icon.dart';
 export 'mini_progress_indicator.dart';
 export 'single_async_scope.dart';
 export 'svg_icon.dart';
+export 'ui_overlay.dart';