+++ magnific_ui/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Change API of Single Async Scope.
+* Replace SvgIcon with SvgAssetIcon, SvgPictureIcon.
+* Add CustomSystemUiOverlay.
+
 ## 0.1.0
 
 * Initial release.